// team08-main/frontend/src/interfaces/UsersInterface.ts
export interface UsersInterface {
    ID?: number;
    UserName?:  		string ;
	Password?:     	string ;
	Email?:     		string ;
	Profile?: 		string;
	ProfileBackground?: 		string;
	FirstName?: 		string;
	LastName?: 			string;
	Age?: 		number;
	Tel?: 			string ;
	Status?: 			string ;
	Point?: number
	QuotaOrder? :number
	QuotaBooking? :number
	PointRate? :number
    DiscountRate? :number
}